Control.Natural

Description

A data type and class for natural transformations.

Newtype for a Natural Transformation

newtype f :~> g infixr 0

A natural transformation suitable for storing in a container.

Constructors

NT

Fields

($$) :: f ~> g

Type Synonym for a Natural Transformation

type (~>) f g = forall x. f x -> g x infixr 0

A natural transformation from f to g.

Conversion functions between the newtype and the synonym

wrapNT :: (forall a. f a -> g a) -> f :~> g

wrapNT builds our natural transformation abstraction out of a natural transformation function.

An alias to NT provided for symmetry with unwrapNT.

unwrapNT :: Transformation f g t => t -> forall a. f a -> g a

unwrapNT is the nonfix version of #. It is used to break natural transformation wrappers, including :~>.

Class for Natural Transformations

class Transformation f g t | t -> f g where

A (natural) transformation is inside t, and contains f and g (typically Functors).

The order of arguments allows the use of GeneralizedNewtypeDeriving to wrap a :~>, but maintain the Transformation constraint. Thus, # can be used on abstract data types.

Minimal complete definition

(#)

Methods

(#) :: t -> forall a. f a -> g a infix 0

The invocation method for a natural transformation.