Control.Object

Description

An Object type, which is a natural transformation into the IO monad.

Documentation

newtype Object f

An Object is a natural transformation from a given Functor f, to IO.

Constructors

Object (f ~> IO)

(#) :: Transformation f g t => t -> forall a. f a -> g a

The invocation method for a natural transformation.